backend/requirements.txt
# Python dependencies for Warehouse Scale Automation Backend

# Flask core
Flask>=3.0.0
flask-smorest>=0.44.0
flask-cors>=4.0.0

# Database
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Schema validation
marshmallow>=3.20.0
marshmallow-sqlalchemy>=0.29.0

# Environment
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
